Refactor shared/mmdb with main/worker separation in mind
