+++ src/main/pluginsServer.ts
@@ -7,11 +7,11 @@ import net, { AddressInfo } from 'net'
 import * as schedule from 'node-schedule'
 
 import { defaultConfig } from '../shared/config'
-import { createMmdbServer, performMmdbStalenessCheck, prepareMmdb } from '../shared/mmdb'
 import { createServer } from '../shared/server'
 import { status } from '../shared/status'
 import { createRedis, delay } from '../shared/utils'
 import { PluginsServer, PluginsServerConfig, Queue, ScheduleControl } from '../types'
+import { createMmdbServer, performMmdbStalenessCheck, prepareMmdb } from './mmdb'
 import { startQueue } from './queue'
 import { startSchedule } from './services/schedule'
 import { startFastifyInstance, stopFastifyInstance } from './web/server'
